src_py/requirements.txt
dstack-sdk==0.1.2
asyncio>=3.4.3
fastapi==0.115.4
uvicorn==0.32.0
openai==1.53.0
python-dotenv==1.0.1
httpx==0.27.2
web3==7.4.0
loguru==0.7.2
pandas==2.2.3
openpyxl==3.1.5